refactor: 결제 마감 배치 작업 파일 이동
