+++ src/main/java/com/meetplus/batch/PaymentCancelSchedule.java
@@ -1,13 +1,9 @@
-package com.meetplus.batch.config;
+package com.meetplus.batch;
 
-import java.util.HashMap;
-import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.Job;
-import org.springframework.batch.core.JobParameter;
 import org.springframework.batch.core.JobParameters;
 import org.springframework.batch.core.JobParametersBuilder;
-import org.springframework.batch.core.launch.JobInstanceAlreadyExistsException;
 import org.springframework.batch.core.launch.JobLauncher;
 import org.springframework.batch.core.repository.JobExecutionAlreadyRunningException;
 import org.springframework.batch.core.repository.JobInstanceAlreadyCompleteException;
@@ -25,9 +21,8 @@ public class PaymentCancelSchedule {
 
     @Scheduled(cron = "0 0 2 * * ?")
     public void runJob() throws Exception {
-//        jobLauncher.run(updatePaymentStatusJob, new JobParametersBuilder().toJobParameters());
         JobParameters jobParameters = new JobParametersBuilder()
-            .addString("time", String.valueOf(System.currentTimeMillis()))  // 고유한 파라미터 추가
+            .addString("paymentCancelTime", String.valueOf(System.currentTimeMillis()))  // 고유한 파라미터 추가
             .toJobParameters();
 
         try {
